+++ databases/embedjs-lmdb/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.11 (2024-10-27)
+
+This was a version bump only for embedjs-lmdb to align it with other projects, there were no code changes.
+
 ## 0.1.10 (2024-10-25)
 
 This was a version bump only for embedjs-lmdb to align it with other projects, there were no code changes.